feat: Refactored some of the HTML parsing code for the leaderboard
Refactored some of the HTML parsing code for the leaderboard. Using css
selectors and xpaths, I'm hoping there will be a performance increase in
comparison to searching the entire DOM, assuming the functions are
working as I think they do. For scraping, all that's left is to write
data for a single leaderboard page to file, update the current page
value in the database, and then scrape the next (updated) current page.

Right now there's a break statement ensuring the scraper stops after 1
loop. Therefore, looping over leaderboard pages AND divisions also needs
to be tested.