feat: Ensured data visualization and scorer module work post-update
Ensured data visualization and scorer module work post-update. The
leaderboard scoring was actually working fine and as expected yesterday,
but it was BECAUSE of the data collection error in "for time" workouts
that was reflected in the leaderboard mis-ranking athletes who DNF'd on
"for time" metcons 18.2, 18.3, and 18.4 that the aggregation error was
seen.