bug: Fixed database schema issue for 18.3
Fixed database schema issue for 18.3. This required no updates to data
collection logic (although the data collection notebook DOES house the
database schema setup). 18.3 was labeled as "leaderboard_18_3_reps" in
the database, when it was supposed to be "leaderboard_18_3_time_secs".
This cause scoring issues in the scorer.py module.