fix: Fixed bug where fight gone bad scores were all dropped
All dropped fight gone bad scores here were recorded to have a score of
-2. These scores (moving forward) WILL be recorded for athletes with
valid entries, but not for those already scraped (which are those who
perform the best, sigh). If fight gone bad is deemed as super
influential in the analysis process, we can go back and rescrape and
update these scores specifically.